requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.2.1 
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0 